+++ src/dipdup/datasources/tezos_tzkt.py
@@ -8,7 +8,6 @@ from collections.abc import Awaitable
 from collections.abc import Callable
 from collections.abc import Generator
 from collections.abc import Sequence
-from dataclasses import fields
 from enum import Enum
 from functools import partial
 from typing import Any
@@ -96,7 +95,7 @@ BIGMAP_FIELDS = (
     'keyType',
     'valueType',
 )
-TZKT_TOKEN_TRANSFER_DATA_FIELDS = (
+TOKEN_TRANSFER_FIELDS = (
     'token',
     'from',
     'to',
@@ -107,6 +106,16 @@ TZKT_TOKEN_TRANSFER_DATA_FIELDS = (
     'transactionId',
     'originationId',
     'migrationId',
+)
+EVENT_FIELDS = (
+    'id',
+    'level',
+    'timestamp',
+    'tag',
+    'payload',
+    'contract',
+    'codeHash',
+    'transactionId',
 )
 
 
@@ -845,7 +854,7 @@ class TzktDatasource(IndexDatasource[TzktDatasourceConfig]):
             last_level,
             offset=offset or 0,
             limit=limit,
-            select=TZKT_TOKEN_TRANSFER_DATA_FIELDS,
+            select=TOKEN_TRANSFER_FIELDS,
             values=True,
             cursor=True,
             **{
@@ -889,20 +898,24 @@ class TzktDatasource(IndexDatasource[TzktDatasourceConfig]):
         offset: int | None = None,
         limit: int | None = None,
     ) -> tuple[TzktEventData, ...]:
-        # FIXME: no tests for function
+        params = self._get_request_params(
+            first_level,
+            last_level,
+            offset=offset or 0,
+            limit=limit,
+            select=EVENT_FIELDS,
+            values=True,
+            cursor=True,
+            **{
+                'contract.in': ','.join(addresses),
+                'tag.in': ','.join(tags),
+            },
+        )
         offset, limit = offset or 0, limit or self.request_limit
         raw_events = await self._request_values_dict(
             'get',
             url='v1/contracts/events',
-            params={
-                'contract.in': ','.join(addresses),
-                'tag.in': ','.join(tags),
-                'level.ge': first_level,
-                'level.le': last_level,
-                'select.values': ','.join(f.name for f in fields(TzktEventData)),
-                'offset.cr': offset,
-                'limit': limit,
-            },
+            params=params,
         )
         return tuple(TzktEventData.from_json(e) for e in raw_events)