+++ CHANGELOG.md
@@ -10,8 +10,9 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 - cli: Import package submodules before starting indexing to fail early on import errors.
 - cli: Fixed ordering of crash reports in `report` group commands.
-- tezos.tzkt.operations: Fixed parsing operations with empty parameters.
+- tezos.tzkt: Fixed parsing contract event data.
 - tezos.tzkt: Fixed issue with processing rollbacks while sync is in progress.
+- tezos.tzkt.operations: Fixed parsing operations with empty parameters.
 
 ## [7.0.0rc5] - 2023-09-06